Merge.pm: fix RT#81156 - behavior name bug
Fix for the bug that you can specify a behavior name in lowercase
but cannot set it later because of a uc() call in set_behavior.
The uc() call has been removed and it is ensured that you can
only use behavior names which are unique in uppercase.

Reported by Matt D. Garrett

Signed-off-by: Stefan Hermes <sh@riseup.net>